Созданы папки для компонентов и утилит, создана функция usePhoneBook
