+++ phone-book/src/utils/getId.js
@@ -1,0 +1,4 @@
+export function getId() {
+    const id = Date.now()
+    return id
+}
